CSR added for the tests and graph views and OMG. No wegihts yet though

+++ src/common/ids.rs
@@ -1,24 +1,24 @@
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub struct DeclId(pub u32);
+pub struct DeclId(pub usize);
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct NodeId(pub u32);
+pub struct NodeId(pub usize);
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct EdgeId(pub u32);
+pub struct EdgeId(pub usize);
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-pub struct HyperArcId(pub u32);
+pub struct HyperArcId(pub usize);
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct SymId(pub u32);
 
 impl DeclId {
-    pub const NONE: Self = Self(u32::MAX);
+    pub const NONE: Self = Self(usize::MAX);
 
     #[inline(always)]
-    pub fn is_none(self) -> bool { self.0 == u32::MAX }
+    pub fn is_none(self) -> bool { self.0 == usize::MAX }
 
     #[inline(always)]
-    pub fn is_some(self) -> bool { self.0 != u32::MAX }
+    pub fn is_some(self) -> bool { self.0 != usize::MAX }
 }
